requirements.txt
pandas
numpy
akshare>=1.14.0
pyyaml
